requirements.txt
psycopg2-binary==2.9.9
confluent-kafka==2.3.0
minio==7.2.16
PyYAML==6.0.1